Move logics to hal
